+++ intelmq/tests/bots/parsers/generic/test_parser_csv.py
@@ -10,6 +10,9 @@ from intelmq.bots.parsers.generic.parser_csv import GenericCsvParserBot
 with open(os.path.join(os.path.dirname(__file__), 'sample_report.csv')) as handle:
     SAMPLE_FILE = handle.read()
 SAMPLE_SPLIT = SAMPLE_FILE.splitlines()
+with open(os.path.join(os.path.dirname(__file__), 'compose_fields.csv')) as handle:
+    COMPOSE_FILE = handle.read()
+COMPOSE_SPLIT = COMPOSE_FILE.splitlines()
 
 EXAMPLE_REPORT = {"feed.name": "Sample CSV Feed",
                   "feed.url": "http://www.samplecsvthreatfeed.com/list",
@@ -34,6 +37,20 @@ EXAMPLE_EVENT2 = EXAMPLE_EVENT.copy()
 EXAMPLE_EVENT2['time.source'] = "2016-12-14T04:19:00+00:00"
 EXAMPLE_EVENT2['source.ip'] = "198.105.221.161"
 EXAMPLE_EVENT2["raw"] = utils.base64_encode(SAMPLE_SPLIT[2].replace('\t', ',')+'\r\n')
+COMPOSE_REPORT = {"feed.name": "Sample CSV Feed",
+                  "raw": utils.base64_encode(COMPOSE_FILE),
+                  "__type": "Report",
+                  "time.observation": "2015-01-01T00:00:00+00:00",
+                  }
+COMPOSE_EVENT = {"feed.name": "Sample CSV Feed",
+                 "__type": "Event",
+                 "source.urlpath": "/foo/",
+                 "source.url": "http://example.com/foo/",
+                 "source.fqdn": "example.com",
+                 "classification.type": "malware",
+                 "raw": utils.base64_encode(COMPOSE_SPLIT[1]+'\r\n'),
+                 "time.observation": "2015-01-01T00:00:00+00:00",
+                 }
 
 
 class TestGenericCsvParserBot(test.BotTestCase, unittest.TestCase):
@@ -72,6 +89,13 @@ class TestGenericCsvParserBot(test.BotTestCase, unittest.TestCase):
         self.run_bot()
         self.assertMessageEqual(0, EXAMPLE_EVENT)
         self.assertMessageEqual(1, EXAMPLE_EVENT2)
+
+    def test_compose_fields(self):
+        self.input_message = COMPOSE_REPORT
+        self.run_bot(parameters={'columns': ['source.fqdn', 'source.urlpath'],
+                                 'compose_fields': {"source.url": "http://{0}{1}"},
+                                 'delimiter': ','})
+        self.assertMessageEqual(0, COMPOSE_EVENT)
 
 
 if __name__ == '__main__':  # pragma: no cover